pyproject.toml
[project]
name = "anime-recomender"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "chromadb>=1.4.1",
    "langchain>=1.2.7",
    "langchain-community>=0.4.1",
    "langchain-groq>=1.1.1",
    "langchain-huggingface>=1.2.0",
    "pandas>=2.3.3",
    "python-dotenv>=1.2.1",
    "sentence-transformers>=5.2.0",
    "streamlit>=1.53.1",
]
